refactor: Generator 디렉터리 생성
- RandomNumber -> RandomNumberGenerator 명칭 변경
