+++ src/main/java/Generator/NumberGenerator.java
@@ -1,4 +1,4 @@
-package Domain;
+package Generator;
 
 public interface NumberGenerator {
     int generateNumber();
